spat_back/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
distro==1.9.0
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
groq==1.1.2
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
pydantic==2.13.0
pydantic_core==2.46.0
PyJWT==2.12.1
python-decouple==3.8
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
whitenoise==6.12.0
